tg_bot/requirements.txt
pyTelegramBotAPI==4.23.0
python-dotenv==1.0.1
SQLAlchemy~=2.0.36
psycopg2-binary==2.9.10
python-decouple==3.8
